// src/modules/services-sait/enums/sait-url.enums.ts
export enum SaitUrlEnum {
    AUTH = "/auth",
    UPLOAD = "/convenios/costehubpack/upload",
    VALIDATE = "/convenios/costehubpack/validar",
    VALIDATION_RESULT = "/convenios/costehubpack/resultado",
    PROCESS = "/convenios/costehubpack/procesar",
    PROCESS_RESULT = "/convenios/costehubpack/resultado",
    STATUS = "/convenios/costehubpack/estado",
    DELIVERY_STATUS= "/convenios/estadoDelivery"
}